requirements/testing.txt
-r pip.txt

pytest-django==3.9.0
django-dynamic-fixture==3.1.0
pytest-cov==2.8.1
